### 2024-06-12 #6861 @vissarion
Hi @fabian-s, @Pentaonia, @vankesteren, @wcjochem, @jayrobwilliams, @fernandomayer would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
