### 2024-06-12 #6878 @editorialbot
[REVIEW]: SpatialGEV: Fast Bayesian inference for spatial extreme value models in R
**Submitting author:** @meixichen (<a href="http://orcid.org/0000-0003-1012-5352">Meixi Chen</a>)
**Repository:** https://github.com/meixichen/SpatialGEV
**Branch with paper.md** (empty if default branch): joss
**Version:** v3
