### 2024-06-12 #6878 @fabian-s
@meixichen could you try finding/adding the missing DOIs? (see above)
