### 2024-06-12 f2bd59b
Shorten paper and add doi
