+++ paper/paper.bib
@@ -3,7 +3,8 @@
 	author={Chen, M. and Ramezan, R. and Lysy, M.},
 	year={2024},
 	journal={Canadian Journal of Statistics (accepted)},
-	url={https://arxiv.org/abs/2110.07051}
+	url={https://arxiv.org/abs/2110.07051},
+	doi={10.48550/arXiv.2110.07051}
 }
 
 @article{tierney-kadane86,
@@ -12,8 +13,41 @@
   journal={Journal of the American Statistical Association},
   year={1986},
   volume={81},
-  pages={82-86}
+  pages={82-86},
+  doi={10.1080/01621459.1986.10478240}
 }
+
+@article{coles98,
+  author = {Coles, S. G. and Casson, E.},
+  year = {1998},
+  title = {Extreme Value Modelling of Hurricane Wind Speeds},
+  journal = {Structural Safety},
+  volume = {20},
+  pages = {283-296},
+  doi = {10.1111/1467-9876.00189}
+}
+
+
+@article{cooley07,
+  author = {Cooley, D. and Nychka, D. and Naveau, P.},
+  year = {2007},
+  title = {Bayesian Spatial Modeling of Extreme Precipitation Return Levels},
+  journal = {Journal of the American Statistical Association},
+  volume = {102},
+  pages = {824-840},
+  doi = {10.1198/016214506000000780}
+}
+
+ @article{sang-gelfand10,
+  author = {Sang, H. and Gelfand, A. E.},
+  year = {2010},
+  title = {Continuous Spatial Process Models for Spatial Extreme Values},
+  journal = {Journal of Agricultural, Biological, and Environmental Statistics},
+  volume = {15},
+  pages = {49-56},
+  doi = {10.1007/s13253-009-0010-1}
+}
+
 
 @article{wood20,
   title={Inference and computation with generalized additive models and their extensions},
@@ -28,27 +62,10 @@
   title={The {SPDE} approach for {Gaussian} and {non-Gaussian} fields: 10 years and still running},
   author={Lindgren, F. K. and Bolin, D. and Rue, H.},
   journal={Spatial Statistics},
-  year={2021}
+  year={2021},
+  doi={10.1016/j.spasta.2022.100599}
 }
 
-@article{coles98,
-  author = {Coles, S. G. and Casson, E.},
-  year = {1998},
-  title = {Extreme Value Modelling of Hurricane Wind Speeds},
-  journal = {Structural Safety},
-  volume = {20},
-  pages = {283-296}
-}
-
-
-@article{cooley07,
-  author = {Cooley, D. and Nychka, D. and Naveau, P.},
-  year = {2007},
-  title = {Bayesian Spatial Modeling of Extreme Precipitation Return Levels},
-  journal = {Journal of the American Statistical Association},
-  volume = {102},
-  pages = {824-840}
-}
 
 @article{schliep10,
   author = {Schliep, E. M. and Cooley, D. and Sain, S.R. and Hoeting, J.A.},
@@ -56,7 +73,8 @@
   title = {A Comparison Study of Extreme Precipitation from Six Different Regional Climate Models via Spatial Hierarchical Modeling},
   journal = {Extremes},
   volume = {13},
-  pages = {219-239}
+  pages = {219-239},
+  doi={10.1007/s10687-009-0098-2}
 }
 
 
@@ -66,18 +84,9 @@
   title = {Bayesian Hierarchical Modeling of Extreme Hourly Precipitation in {Norway}},
   journal = {Environmetrics},
   volume = {26},
-  pages = {89-106}
+  pages = {89-106},
+  doi={10.1002/env.2301}
 }
-
-@article{tye-cooley15,
[…]
