+++ paper/paper.md
@@ -47,34 +47,7 @@ MathJax.Hub.Config({
 
 # Summary
 
-Extreme weather phenomena such as floods and hurricanes are of great concern due to their potential to cause extensive damage. To develop more reliable damage prevention protocols, statistical models are often used to infer the chance of observing an extreme weather event at a given location [@coles98; @cooley07; @sang-gelfand10]. Here we present **SpatialGEV**, an \textsf{R} package providing a fast and convenient toolset for analyzing spatial extreme values using a hierarchical Bayesian modeling framework. In this framework, the marginal behavior of the extremes is given by a generalized extreme value (GEV) distribution, whereas the spatial dependence between locations is captured by modeling the GEV parameters as spatially varying random effects following a Gaussian process (GP). Model inference is carried out using an efficient implementation of the Laplace approximation, which produces highly accurate posterior estimates several orders of magnitude faster than Markov Chain Monte Carlo (MCMC) methods. Users are provided with a streamlined way to build and fit various GEV-GP models in \textsf{R}, which are compiled in \textsf{C++} under the hood. For downstream analyses, the package offers methods for Bayesian parameter estimation and forecasting of extreme events. 
-
-# Background
-
-Let $y_{ij}$ denote observation $j$ of an extreme weather event at spatial location $i$, of which the two-dimensional spatial coordinates are $\xx_i$. The general form of the GEV-GP models that can be fit using **SpatialGEV** is
-\begin{equation}
-\begin{aligned}
-  y_{ij} &\iid \operatorname{GEV}(a(\xx_i), \exp(b(\xx_i)), s(\xx_i)),\\
-  u(\xx) &\sim \operatorname{GP}(\bm{c}_u(\xx)^T\bm{\beta}_u(\xx), \ \mathcal{K}(\xx,\xx' \mid \bm{\eta}_u)),
-\end{aligned}
-\end{equation}
-where $\operatorname{GEV}(a, b, s)$ denotes a GEV distribution whose cumulative density function (CDF) is given by
-\begin{equation}
-F(y \mid a, b, s) = 
-\begin{cases}
-\exp\left\{-\left(1+s\cdot \frac{y-a}{b}\right)^{-1/s}\right\}, \ &s\neq 0,\\
-\exp\left\{-\exp\left(-\frac{y-a}{b}\right)\right\}, \ &s=0, 
-\end{cases}
-\end{equation}
-with location parameter $a$, positive scale parameter $b$ and shape parameter $s$,
-$\operatorname{GP}(\mu(\xx), \mathcal{K}(\xx,\xx'))$ is a Gaussian process with mean function $\mu(\xx)$ (possibly depending on covariates $\bm{c}(\xx)$ via coefficients $\bm{\beta}(\xx)$) and covariance kernel $\mathcal{K}(\xx,\xx')$, and $u\in\{a,b,s\}$ is any subset of the GEV parameters which we would like to model as spatially varying.
-
-The GEV-GP model has important applications in meteorological studies. For example, let $y=y(\xx)$ denote the amount of rainfall at a spatial location $\xx$. To forecast extreme rainfalls, it is often of interest for meteorologists to estimate the $1/p$-year rainfall return level $z_p(\xx)$, which is the value above which precipitation levels at location $\xx$ occur with probability $p$, i.e.,
-\begin{equation}
-\Pr\big(y(\xx) > z_p(\xx)\big) = 1-F\big(z_p(\xx)\mid a(\xx), b(\xx), s(\xx)\big) = p,
-\label{eq:return-level}
-\end{equation}
[…]
