+++ paper/paper.bib
@@ -55,7 +55,8 @@
   journal={Test},
   year={2020},
   volume={29},
-  pages={307-339}
+  pages={307-339},
+  doi={10.1007/s11749-020-00711-5}
 }
 
 @article{lindgren-etal21,
@@ -95,7 +96,7 @@
   booktitle = {The Handbook of {Markov Chain Monte Carlo}},
   publisher = {Chapman \& Hall / CRC Press},
   year = {2011},
-  chapter = {5},
+  chapter = {5}
 }
 
 @article{hoffman-gelman14,
@@ -125,7 +126,7 @@
   journal = {Journal of Statistical Software},
   volume = {63},
   pages = {1-25},
-  pages = {10.18637/jss.v063.i19}
+  doi = {10.18637/jss.v063.i19}
 }
 
 
