### 2024-06-12 #6878 @meixichen
Hello, I wonder if 1494 word count is still considered too long? Thanks!
