### 2024-06-12 #6878 @fabian-s
**Contribution and authorship:**

Seems like only @meixichen has a meaningful number of commits here -- what are the contributions of the other two authors? The only commit of @mlysy from 4 years ago (https://github.com/meixichen/SpatialGEV/commit/47be58e10ef277a041d2c60817e10252939bad63) seems more like housekeeping to me and, in any case, git blame (https://github.com/meixichen/SpatialGEV/blame/master/inst/include/SpatialGEV/utils.hpp) seems to indicate that very little of the code committed by them back then still remains part of the codebase?
