### 2024-06-12 meixichen/SpatialGEV#13 @fabian-s
JOSS paper examples don't run
Code in your JOSS paper should be executable:

```
> fit <- spatialGEV_fit(y = y, locs = locs, random = "ab",
+                       init_param = list(a = rep(4, n_loc),
+                                         log_b = rep(0,n_loc),
+                                         s = -2,
+                                         beta_a = 4, beta_b = 0,
+                                         log_sigma_a = 0, log_ell_a = 1,
+                                         log_sigma_b = 0, log_ell_b = 1),
+                       reparam_s = "positive", kernel="exp",
+                       X_a = matrix(1, nrow=n_loc, ncol=1),
+                       X_b = matrix(1, nrow=n_loc, ncol=1),
+                       silent=T)
Error in spatialGEV_fit(y = y, locs = locs, random = "ab", init_param = list(a = rep(4,  :
  argument "data" is missing, with no default
  ```

  openjournals/joss-reviews#6878
