### 2024-06-12 meixichen/SpatialGEV#14 @fabian-s
code quality issues 
some issues found by `goodpractice::gp()`, especially test coverage seems way too low.

It is good practice to

  ✖ write unit tests for all functions, and all package code in
    general. 15% of code lines are covered by test cases.

    R/grid_location.R:26:NA
    R/grid_location.R:27:NA
    R/grid_location.R:29:NA
    R/grid_location.R:30:NA
    R/grid_location.R:31:NA
    ... and 482 more lines

  ✖ add a "URL" field to DESCRIPTION. It helps users find
    information about your package online. If your package does not
    have a homepage, add an URL to GitHub, or the CRAN package package
    page.
  ✖ add a "BugReports" field to DESCRIPTION, and point it to a
    bug tracker. Many online code hosting services provide bug trackers
    for free, https://github.com, https://gitlab.com, etc.
  ✖ use '<-' for assignment instead of '='. '<-' is the
    standard, and R users and developers are used it and it is easier
    to read your code for them if you use '<-'.

    tests/testthat/test-model_a_exp.R:25:15
    tests/testthat/test-model_a_matern.R:25:15
    tests/testthat/test-model_ab_exp.R:30:15
    tests/testthat/test-model_ab_matern.R:30:15
    tests/testthat/test-model_abs_exp.R:36:15
    ... and 1 more lines

    vapply() instead.

    data-raw/simulatedData2.R:34:17
    R/r_nll.R:91:24
    R/r_nll.R:108:19
    R/r_nll.R:121:19
    R/spatialGEV_fit.R:265:12

  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...)
    and 1:NCOL(...) expressions. They are error prone and result 1:0 if
    the expression on the right hand side is zero. Use seq_len() or
    seq_along() instead.

    R/spatialGEV_sample.R:74:25
    R/spatialGEV_sample.R:78:25
    R/spatialGEV_sample.R:83:25
    R/spatialGEV_sample.R:100:16

  ✖ not import packages as a whole, as this can cause name
    clashes between the imported packages. Instead, import only the
    specific functions you need.

see  openjournals/joss-reviews#6878
