### 2024-06-12 #6878 @fabian-s
** Automated tests**: test coverage seems to be rather low, see meixichen/SpatialGEV#14
