### 2024-06-12 #6878 @fabian-s
**Functionality**:
Neither the package vignette nor the JOSS paper provide examples or details on how covariate information is included in the model. Are these linear effects? If so, how are the coefficients sampled and based on which priors? Do the covariates need to be scaled/centered?
