### 2024-06-12 #6878 @fabian-s
**Documentation**:
Your github README.md and your vignette use `y` as the first argument to `spatialGEV_fit`, the package and the JOSS paper use `data`, this seems to have diverged.
