### 2024-06-12 meixichen/SpatialGEV#13 @meixichen
> PS:
>
> Your github README.md and your vignette use y as the first argument to spatialGEV_fit, the package and the JOSS paper use data, this seems to have diverged...

Thank you for the note. Just wanted to clarify, I believe our README on the joss branch (which is for review) is up to date, but I know our master branch is not up to date and we plan to merge joss with master later. Is it possible that you were looking at the master branch? Thanks!
