### 2024-06-18 90dda42
Resolve issues raised by goodpractice::gp
