+++ R/SpatialGEV-package.R
@@ -1,6 +1,6 @@
 #' @keywords internal
-#' @import mvtnorm
-#' @import evd
+#' @importFrom mvtnorm rmvnorm dmvnorm
+#' @importFrom evd dgev rgev qgev
 #' @importFrom stats dist nlminb rnorm runif rexp setNames quantile
 #' @importFrom methods as is
 #' @rawNamespace useDynLib(SpatialGEV, .registration=TRUE);useDynLib(SpatialGEV_TMBExports)
