+++ NAMESPACE
@@ -15,10 +15,13 @@ export(spatialGEV_fit)
 export(spatialGEV_model)
 export(spatialGEV_predict)
 export(spatialGEV_sample)
-import(evd)
-import(mvtnorm)
+importFrom(evd,dgev)
+importFrom(evd,qgev)
+importFrom(evd,rgev)
 importFrom(methods,as)
 importFrom(methods,is)
+importFrom(mvtnorm,dmvnorm)
+importFrom(mvtnorm,rmvnorm)
 importFrom(stats,dist)
 importFrom(stats,nlminb)
 importFrom(stats,quantile)
