### 2024-06-18 meixichen/SpatialGEV#14 @meixichen
Thanks a lot for raising these points. I ran `goodpractice::gp()` on the joss branch again after fixing some of these issues, now I only have
```
It is good practice to

    data-raw/CA-snowdata.R:25:81
    data-raw/simulatedData.R:18:81
    data-raw/simulatedData.R:27:81
    data-raw/simulatedData.R:29:81
    R/data.R:3:81
    ... and 347 more lines
```
