### 2024-06-19 15b74f5
Add methods for formula interface and tidy (tidybayes / broom).
