+++ R/interface.R
@@ -1,0 +1,175 @@
+#' @export
+spatial_gev <- function(x, ...){
+  UseMethod("spatial_gev")
+}
+
+#' @export
+spatial_gev.formula <- function(
+    formula, data, coordinates = ~ x + y,
+    method = c("laplace", "maxsmooth"),
+    kernel = c("spde", "matern", "exp"),
+    reparam_s = c("zero", "unconstrained", "negative", "positive"),
+    prior = list(),
+    control = list(),
+    init = NULL
+){
+  method <- match.arg(method)
+  kernel <- match.arg(kernel)
+  
+  formula <- Formula::Formula(formula)
+  formula <- expand_dot(formula)
+  flen <- length(formula)
+  stopifnot(identical(flen[1], 1L), flen[2] <= 3 & flen[2])
+  
+  lhs <- rlang::f_lhs(formula)
+  data <- tidyr::chop(data, lhs) # make data nloc x nvar with a list response
+  
+  Xmats <- regression_matrix(formula, data)
+  
+  ret <- list(
+    data = get(lhs, data),
+    locs = model.frame(coordinates, data),
+    random  = switch(flen[2], "a", "ab", "abc"),
+    method = method,
+    reparam_s = reparam_s,
+    kernel = kernel,
+    formula = formula,
+    control = set_control(control),
+    prior   = set_prior(prior)
+  )
+  if(is.null(init)){
+    ret$init_param <- init_params(
+      x = ret$data,
+      random = Filter(Negate(is.null), lapply(Xmats, ncol)),
+      kernel = ret$kernel,
+      use_fgev = TRUE
+    )
+  }
+  ret <- c(ret, setNames(Xmats, paste0("X_", names(Xmats))))
+  mycall <- rlang::call2("spatialGEV_fit", !!!ret[c(1:6,9:12)], !!!ret$control, !!!ret$prior, .ns = "SpatialGEV")
+  fit <- eval(mycall)
+  attr(fit, "formula") <- ret$formula
+  attr(fit, "call") <- mycall
+  fit
+}
+
+set_prior <- function(prior){
+  ret <- list(
+    nu              = 1,
+    s_prior         = NULL,
+    beta_prior      = NULL,
+    matern_pc_prior = NULL
+  )
+  nms <- names(ret)
+  nms_set <- names(prior)
+  ret[nms_set] <- prior
+  if (length(noNms <- nms_set[!nms_set %in% nms])){
+    stop("Unknown control parameters passed: ", paste(noNms, collapse = ", "))
+  }
+  ret
+}
+
+set_control <- function(control){
+  ret <- list(
+    return_levels         = 0,
+    get_return_levels_cov = TRUE,
+    sp_thres              = -1,
+    adfun_only            = FALSE,
+    ignore_random         = FALSE,
+    get_hessian           = TRUE,
+    silent                = TRUE,
+    mesh_extra_init = list(
+      a     = 0,
+      log_b = -1,
+      s     = 0.001
+    )
+  )
+  nms <- names(ret)
+  nms_set <- names(control)
+  ret[nms_set] <- control
+  if (length(noNms <- nms_set[!nms_set %in% nms])){
+    stop("Unknown control parameters passed: ", paste(noNms, collapse = ", "))
+  }
+  ret
+}
+
+#' @export
+init_params <- function(x, random, kernel, use_fgev = TRUE){
+  # For each parameter need:
+  # list("a", "log_b", "s") where length is 1 when not random and nloc when random
+  # For each random parameter above, need
+  hyper <- if(identical(kernel, "exp")){
+    c("log_sigma", "log_ell")
+  } else {
+    c("log_sigma", "log_kappa")
+  }
[…]
