### 2024-06-19 meixichen/SpatialGEV#14 @fabian-s
hmm, I don't see it?

```
10:54 $ git switch joss-devel
Switched to branch 'joss-devel'
(base) ✔ ~/reviews/JOSS/SpatialGEV [joss-devel L|✔]
10:55 $ R

R version 4.4.0 (2024-04-24) -- "Puppy Cup"
[...]

> goodpractice::gp()
Preparing: covr
Preparing: cyclocomp
Preparing: description
Preparing: lintr
  |======================================================================| 100%
Preparing: namespace
Preparing: rcmdcheck
── GP SpatialGEV ───────────────────────────────────────────────────────────────

It is good practice to

  ✖ write unit tests for all functions, and all package code in
    general. 18% of code lines are covered by test cases.

    R/grid_location.R:22:NA
    R/grid_location.R:23:NA
    R/grid_location.R:25:NA
    R/grid_location.R:26:NA
    R/grid_location.R:27:NA
    ... and 542 more lines

    data-raw/CA-snowdata.R:25:81
    data-raw/simulatedData.R:18:81
    data-raw/simulatedData.R:27:81
    data-raw/simulatedData.R:29:81
    R/data.R:3:81
    ... and 348 more lines

  ✖ omit trailing semicolons from code lines. They are not
    needed and most R coding standards forbid them

    tests/dontrun/test-spatialGEV_sample.R:7:18

    vapply() instead.

    data-raw/CA-snowdata.R:65:25
    data-raw/simulatedData2.R:34:17
    R/old/spatialGEV_fit.R:265:12
    R/spatialGEV_model.R:92:13
    R/spatialGEV_model.R:98:14
    ... and 6 more lines

  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...)
    and 1:NCOL(...) expressions. They are error prone and result 1:0 if
    the expression on the right hand side is zero. Use seq_len() or
    seq_along() instead.

    inst/include/SpatialGEV/templates/make_templates.R:49:11
    R/old/spatialGEV_sample.R:43:25
    R/old/spatialGEV_sample.R:46:25
    R/old/spatialGEV_sample.R:51:25
    R/old/spatialGEV_sample.R:68:16

  ✖ not import packages as a whole, as this can cause name
    clashes between the imported packages. Instead, import only the
    specific functions you need.
```

this is at commit 90dda42 ("resolve issues raised by ...")

I would also expect you to re-unite diverged branches (at least: master, joss, joss-dev) so that the JOSS publication refers to a (tagged release of the) package that is not some dead end in your development history, and for your JOSS publication to accurately reflect the features etc of that version.
