### 2024-06-19 meixichen/SpatialGEV#14 @fabian-s
Can't confirm -- i still see:

```
✖ write unit tests for all functions, and all package code in
    general. 18% of code lines are covered by test cases.

    R/grid_location.R:22:NA
    R/grid_location.R:23:NA
    R/grid_location.R:25:NA
    R/grid_location.R:26:NA
    R/grid_location.R:27:NA
    ... and 543 more lines
```

which seems way too low in terms of test coverage.

I would also like to see you merge all these diverged branches back into master for the JOSS publication so that the JOSS publication does not refer to a dead end in your development history but instead a tagged and accessible release version (or: so that the only divergence is paper/paper.md etc, not anything in your R package code).
