### 2024-06-19 34023b0
Add extra repo to test coverage workflow
