+++ .github/workflows/test-coverage.yaml
@@ -21,6 +21,11 @@ jobs:
         with:
           use-public-rspm: true
 
+      - name: Add inla repo
+        run: |
+          writeLines('options(repos = c(INLA = "https://inla.r-inla-download.org/R/stable/", getOption(\"repos\")))', file("~/.Rprofile", "ab"))
+        shell: Rscript {0}
+
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
           extra-packages: covr
