### 2024-06-20 ec985cd
Add codecov badge to README.md
