### 2024-06-20 b9ae81c
Add checks and unittests for grid_location
