### 2024-06-20 896d0c3
Add more checks and unittests for matern_pc_prior
