### 2024-06-20 e4c3f01
Add unittests for model fit/sample/predict functions
