### 2024-06-20 a58c741
Add logs to argument of model fit unittest helper function
