### 2024-06-21 52c0f9b
Fix predict method for model a with only 1 test location, and add unittests for model a
