+++ tests/testthat/test-spatialGEV_fit.R
@@ -8,61 +8,64 @@ y <- simulatedData$y[1:n_loc]
 locs <- simulatedData$locs[1:n_loc,]
 beta_a <- mean(a)
 beta_b <- mean(logb)
+rl_list <- list(0, c(0.5, 0.9))
 
 test_that("Test that `spatialGEV_fit` runs without error for Matern/SPDE.", {
-  rls <- c(0.5, 0.9)
-  for (kernel in c("matern", "spde")){
-    # Model a
-    fit_a <- spatialGEV_fit(
-      data = y,
-      locs = locs,
-      random = "a",
-      init_param = list(
-        beta_a = beta_a,
-        a = rep(0, n_loc),
-        log_b = 0,
-        s = 0,
-        log_sigma_a = 0,
-        log_kappa_a = 0
-      ),
-      reparam_s = "zero",
-      kernel = kernel,
-      return_levels = rls,
-      silent = TRUE
-    )
-    test_model_fit(fit_a, n_loc, n_random=1, n_rl=length(rls))
-    # Model ab
-    fit_ab <- spatialGEV_fit(
-      data = y,
-      locs = locs,
-      random = "ab",
-      init_param = list(
-        beta_a = beta_a,
-        beta_b = beta_b,
-        a = rep(0, n_loc),
-        log_b = rep(0, n_loc),
-        s = 0,
-        log_sigma_a = 0,
-        log_kappa_a = 0,
-        log_sigma_b = 0,
-        log_kappa_b = 0
-      ),
-      reparam_s = "positive",
-      kernel = kernel,
-      beta_prior = list(
-        beta_a=c(0,100),
-        beta_b=c(0,10)
-      ),
-      X_a = matrix(1, nrow=n_loc, ncol=1),
-      X_b = matrix(1, nrow=n_loc, ncol=1),
-      matern_pc_prior = list(
-        matern_a=matern_pc_prior(1e5,0.95,5,0.1),
-        matern_b=matern_pc_prior(1e5,0.95,3,0.1)
-      ),
-      return_levels = rls,
-      silent = TRUE
-    )
-    test_model_fit(fit_ab, n_loc, s_true=logs, n_random=2, n_rl=length(rls))
+  for (rls in rl_list){
+    for (kernel in c("matern", "spde")){
+      # Model a
+      fit_a <- spatialGEV_fit(
+        data = y,
+        locs = locs,
+        random = "a",
+        init_param = list(
+          beta_a = beta_a,
+          a = rep(0, n_loc),
+          log_b = 0,
+          s = 0,
+          log_sigma_a = 0,
+          log_kappa_a = 0
+        ),
+        reparam_s = "zero",
+        kernel = kernel,
+        return_levels = rls,
+        silent = TRUE
+      )
+      n_rl <- ifelse(rls[1]==0, 0, length(rls))
+      test_model_fit(fit_a, n_loc, n_random=1, n_rl=n_rl)
+      # Model ab
+      fit_ab <- spatialGEV_fit(
+        data = y,
+        locs = locs,
+        random = "ab",
+        init_param = list(
+          beta_a = beta_a,
+          beta_b = beta_b,
+          a = rep(0, n_loc),
+          log_b = rep(0, n_loc),
+          s = 0,
+          log_sigma_a = 0,
+          log_kappa_a = 0,
+          log_sigma_b = 0,
+          log_kappa_b = 0
+        ),
+        reparam_s = "positive",
+        kernel = kernel,
+        beta_prior = list(
+          beta_a=c(0,100),
+          beta_b=c(0,10)
+        ),
+        X_a = matrix(1, nrow=n_loc, ncol=1),
+        X_b = matrix(1, nrow=n_loc, ncol=1),
+        matern_pc_prior = list(
[…]
