### 2024-06-24 33aea2b
Add unittest case for predicting at n>1 locations
