### 2024-06-24 8d1f136
Add test for sim_cond_normal
