### 2024-06-24 2cb2131
Add unittests for model abs and sim_cond_nrmal
