### 2024-06-24 e0632f8
Reduce the number of tests for nll checks to speed up devtools::test()
