### 2024-06-24 acae513
Remove one redundant test case
