### 2024-06-24 #6878 @meixichen
> @meixichen any news re. my questions/requests above?

@fabian-s Regarding https://github.com/meixichen/SpatialGEV/issues/14 I have merged joss with master and added more unittests. Test coverage is close to  90% now https://app.codecov.io/github/meixichen/SpatialGEV/tree/master

I am still working on the other requests and will reply here after I am done with them. For authorship, many conversations about functionalities and code design were done off-line or over emails. Some of mlysy's code were in the joss branch and were not reflected in master, but now after merging the joss into master we should see it.
