### 2024-06-24 meixichen/SpatialGEV#14 @meixichen
Added more unittests and merged joss into master. Now we have a codecov badge in README and test coverage is 89.5% now: https://app.codecov.io/github/meixichen/SpatialGEV.
