+++ CONTRIBUTING.md
@@ -1,0 +1,27 @@
+# Contributing to SpatialGEV
+
+Thank you very much for your interest in SpatialGEV! Please see below if you 
+have questions about the package or would like to contribute to this project.
+
+## Have a Question or a Bug?
+
+Please take a look at our [documentation](https://cran.r-project.org/web/packages/SpatialGEV/vignettes/SpatialGEV-vignette.html) and existing [issues](https://github.com/meixichen/SpatialGEV/issues) if you have not. 
+If you could not find an answer there, please open an [Issue](https://github.com/meixichen/SpatialGEV/issues/new) 
+and describe your question/bug with as much context as you can and reproducible code if possible. While we do not
+provide user support at this time, we will try our best to address questions and issues.
+
+## Code Contribution
+
+We are happy that you are interested in contributing to SpatialGEV! 
+If you would like to suggest an enhancement to the package, please first make sure you are using the latest
+version of the package hosted on our GitHub repository. We suggest that you check the [documentation](https://cran.r-project.org/web/packages/SpatialGEV/vignettes/SpatialGEV-vignette.html) and 
+the existing [issues](https://github.com/meixichen/SpatialGEV/issues) to find out whether the functionality 
+already exists or is under development. 
+
+If you believe the suggested enhancement can benefit a wide range of users, you are welcome to submit a new [GitHub issue](https://github.com/meixichen/SpatialGEV/issues) with the following content:
+
+- A descriptive title for the suggestion.
+- Details about the enhancement and how it can benefit most users of SpatialGEV.
+
+If you already have some idea about how to implement the new features, please fork the SpatialGEV repository,
+write and test your code, and submit a [PR](https://docs.github.com/en/get-started/exploring-projects-on-github/contributing-to-a-project#making-a-pull-request), after submitting an issue as described above. 
