### 2024-07-08 4a76d4a
Add simulated data with covariates
