### 2024-07-08 1ee6bbf
Update simulatedData_withcov with covariates
