### 2024-07-08 bb465dc
Add instructions on including covariates in the vignette and delete the 25-year return level in case study for conciseness
