+++ vignettes/SpatialGEV-vignette.Rmd
@@ -48,11 +48,12 @@ where $a\in\mathbb{R}$, $b>0$, and $s\in\mathbb{R}$ are location, scale, and sha
 A Gaussian process $z(\xx)\sim \mathcal{GP}(\mu(\xx_{cov}), k(\xx, \xx'))$ is fully characterized by its mean $\mu(\xx_{cov})$ and kernel function $k(\xx, \xx') = \cov( z(\xx), z(\xx') )$, which captures the strength of the spatial correlation between locations. The mean is a function of parameters $\bm{\beta}$ and covariates $\xx_{cov}=(x_1,\ldots,x_p)'$. We assume that given the locations, the data follow independent GEV distributions each with their own parameters. The complete GEV-GP hierarchical model then becomes
 \begin{equation}
 \begin{aligned}
-y_{ik} \mid a(\xx_i), b(\xx_i), s & \ind \tx{GEV}\big( a(\xx_i), \exp( b(\xx_i) ), \exp(s)\big)\\
+y_{ik} \mid a(\xx_i), b(\xx_i), s & \ind \tx{GEV}\big( a(\xx_i), \exp( b(\xx_i) ), \exp(s(\xx_i))\big)\\
 a(\xx) \mid \bm{\beta}_a, \tth_a &\sim \mathcal{GP}\big( \XX_a\bm{\beta}_a, k(\xx, \xx' \mid \tth_a) \big)\\
 log(b)(\xx) \mid \bm{\beta}_b, \tth_b &\sim \mathcal{GP}\big( \XX_b\bm{\beta}_b, k(\xx, \xx' \mid \tth_b) \big)\\
 s(\xx) \mid \bm{\beta}_s, \tth_s &\sim \mathcal{GP}\big( \XX_s\bm{\beta}_s, k(\xx, \xx' \mid \tth_a) \big).
 \end{aligned}
+\label{eqn:gev-gp-model}
 \end{equation}
 In this package, a uniform prior $\pi(\tth) \propto 1$ is specified on the fixed effect and hyperparameters 
 \[\tth=(s, \bm{\beta}_a, \bm{\beta}_b, \bm{\beta}_s, \tth_a, \tth_b, \tth_s).\] 
@@ -117,7 +118,7 @@ grid_plot(unique(locs$x), unique(locs$y), matrix(logs, ncol=sqrt(n_loc)),
           "Spatial variation of log(s)", y_lab="")
 ```
 
-<img src="sim1-plot-abs-1.png" title="plot of chunk sim1-plot-abs" alt="plot of chunk sim1-plot-abs" width="100%" />
+<img src="figure/sim1-plot-abs-1.png" alt="plot of chunk sim1-plot-abs" width="100%" />
 
 Number of observations at each location is shown in the figure below.
 
@@ -127,7 +128,7 @@ barplot(sapply(y, length),
         main = "Summary of number of observations per location")
 ```
 
-<img src="sim1-plot-num-obs-per-loc-1.png" alt="plot of chunk sim1-plot-num-obs-per-loc" width="100%" />
+<img src="figure/sim1-plot-num-obs-per-loc-1.png" alt="plot of chunk sim1-plot-num-obs-per-loc" width="100%" />
 
 Below are histograms of observations at $8$ randomly sampled locations.
 
@@ -143,7 +144,7 @@ for (i in sam_inds){
 }
 ```
 
-<img src="sim-1plot-y-1.png" alt="plot of chunk sim-1plot-y" width="100%" />
+<img src="figure/sim-1plot-y-1.png" alt="plot of chunk sim-1plot-y" width="100%" />
 
 ### Model fitting
[…]
