### 2024-07-08 cffc46f
Add doc for simulatedData with covariates
