+++ man/simulatedData_withcov.Rd
@@ -1,0 +1,33 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{simulatedData_withcov}
+\alias{simulatedData_withcov}
+\title{Simulated dataset 1 with covariates}
+\format{
+A list containing the simulation parameters and simulated data on a 20x20 grid:
+\describe{
+\item{locs}{A 400x2 matrix. First column contains longitudes and second contains latitudes}
+\item{a}{A length 400 vector. GEV location parameters}
+\item{logb}{A length 400 vector. Log-transformed GEV scale parameters}
+\item{logs}{A scalar. Log-transformed GEV shape parameter shared across space}
+\item{y}{A length 400 list of vectors which are observations simulated at
+each location}
+\item{beta_a}{A length 3 vector containing the regression coefficients for
+a, with the first element being the intercept, the second element as the
+coefficient for covariate 1, and the third for covariate 2}
+\item{beta_b}{A length 2 vector containing the regression coefficients for
+b, with the first element being the intercept and the second element as the
+coefficient for covariate 1}
+\item{covariates}{A 400x3 design matrix containing the covariates, with the
+first column being all 1s for intercept}
+}
+}
+\usage{
+simulatedData_withcov
+}
+\description{
+A list of data used for package testing and demos. Both \code{a} and \code{logb} are simulated on smooth
+deterministic surfaces depending linearly on two covariates.
+}
+\keyword{datasets}
