### 2024-07-08 94fabd1
Add max step to spatialgev_fit and add tests
