### 2024-07-08 2203082
Add utils::tail to namespace and add contributing.md to rbuildignore
