+++ R/SpatialGEV-package.R
@@ -3,6 +3,7 @@
 #' @importFrom evd dgev rgev qgev
 #' @importFrom stats dist nlminb rnorm runif rexp setNames quantile
 #' @importFrom methods as is
+#' @importFrom utils tail
 #' @rawNamespace useDynLib(SpatialGEV, .registration=TRUE);useDynLib(SpatialGEV_TMBExports)
 "_PACKAGE"
 
