+++ NAMESPACE
@@ -29,4 +29,5 @@ importFrom(stats,rexp)
 importFrom(stats,rnorm)
 importFrom(stats,runif)
 importFrom(stats,setNames)
+importFrom(utils,tail)
 useDynLib(SpatialGEV, .registration=TRUE);useDynLib(SpatialGEV_TMBExports)
