### 2024-07-09 20ca60e
Still allow provding a list of est and var to data argument for method=maxsmooth
