### 2024-07-09 c0ac777
Add default s prior for maxsmooth approach and checks data argument input before calculating max step
