### 2024-07-09 61466b5
Add unittests for model abs case using exp kernel and modify other tests
