+++ tests/testthat/test-spatialGEV_fit.R
@@ -16,8 +16,7 @@ test_that("Test that `spatialGEV_fit` runs without error for Matern/SPDE.", {
   for (ii in seq_along(test_settings[,1])){
     kernel <- as.character(test_settings$kernel)[ii]
     rls <- test_settings$rl[[ii]]
-    # Model a
-    fit_a <- spatialGEV_fit(
+    fit_args <- list(
       data = y,
       locs = locs,
       random = "a",
@@ -34,39 +33,23 @@ test_that("Test that `spatialGEV_fit` runs without error for Matern/SPDE.", {
       return_levels = rls,
       silent = TRUE
     )
+    # Model a
+    fit_a <- do.call(spatialGEV_fit, fit_args)
     n_rl <- ifelse(rls[1]==0, 0, length(rls))
     test_model_fit(fit_a, n_loc, n_random=1, n_rl=n_rl)
     # Model ab
-    fit_ab <- spatialGEV_fit(
-      data = y,
-      locs = locs,
-      random = "ab",
-      init_param = list(
-        beta_a = beta_a,
-        beta_b = beta_b,
-        a = rep(0, n_loc),
-        log_b = rep(0, n_loc),
-        s = 0,
-        log_sigma_a = 0,
-        log_kappa_a = 0,
-        log_sigma_b = 0,
-        log_kappa_b = 0
-      ),
-      reparam_s = "positive",
-      kernel = kernel,
-      beta_prior = list(
-        beta_a=c(0,100),
-        beta_b=c(0,10)
-      ),
-      X_a = matrix(1, nrow=n_loc, ncol=1),
-      X_b = matrix(1, nrow=n_loc, ncol=1),
-      matern_pc_prior = list(
-        matern_a=matern_pc_prior(1e5,0.95,5,0.1),
-        matern_b=matern_pc_prior(1e5,0.95,3,0.1)
-      ),
-      return_levels = rls,
-      silent = TRUE
-    )
+    fit_args$random <- "ab"
+    fit_args$init_param <- list(beta_a = beta_a, beta_b = beta_b,
+                                a = rep(0, n_loc), log_b = rep(0, n_loc),  
+                                s = 0,
+                                log_sigma_a = 0, log_kappa_a = 0,
+                                log_sigma_b = 0, log_kappa_b = 0)
+    fit_args$reparam_s <- "positive"
+    fit_args$beta_prior <- list(beta_a=c(0,100), beta_b=c(0,10))
+    fit_args$X_a <- fit_args$X_b <- matrix(1, nrow=n_loc, ncol=1)
+    fit_args$matern_pc_prior <- list(matern_a=matern_pc_prior(1e5,0.95,5,0.1),
+                                     matern_b=matern_pc_prior(1e5,0.95,3,0.1))
+    fit_ab <- do.call(spatialGEV_fit, fit_args)
     test_model_fit(fit_ab, n_loc, s_true=logs, n_random=2, n_rl=n_rl)
   }
 })
@@ -74,7 +57,7 @@ test_that("Test that `spatialGEV_fit` runs without error for Matern/SPDE.", {
 test_that("Test that `spatialGEV_fit` runs without error for Expo kernel.", {
   rls <- 0.9
   # Model a
-  fit_a <- spatialGEV_fit(
+  fit_args <- list(
     data = y,
     locs = locs,
     random = "a",
@@ -94,49 +77,30 @@ test_that("Test that `spatialGEV_fit` runs without error for Expo kernel.", {
     return_levels = rls,
     silent = TRUE
   )
+  fit_a <- do.call(spatialGEV_fit, fit_args)
   test_model_fit(fit_a, n_loc, n_random=1, n_rl=length(rls))
[…]
