### 2024-07-10 #6878 @vissarion
Hi @fernandomayer just checking if there are any issues that blocks you to start with your review.
