### 2024-07-10 meixichen/SpatialGEV#14 @fabian-s
Thanks, this fixes all the issues I reported earlier. I'm opening another issue (sorry) for additional problems I found when trying to install the package and rebuild your vignette...
