### 2024-07-10 meixichen/SpatialGEV#22 @fabian-s
Does not pass R CMD check 
- `DESCRIPTION`:  Your `Suggests:` should include all packages like `fields` and `evd` that are used in your vignettes

- Vignette does not currently include a properly executed example for including covariates:
https://github.com/meixichen/SpatialGEV/blob/61466b5dde6e5f36965f2244fd8034224076cac3/vignettes/SpatialGEV-vignette.Rmd#L314C1-L340C1 is all just error messages?

-  `R CMD check` --as-cran does not pass:

```
* checking CRAN incoming feasibility ... [7s/15s] WARNING
Maintainer: ‘Meixi Chen <meixi.chen@uwaterloo.ca>’
Insufficient package version (submitted: 1.0.1, existing: 1.0.1)
Suggests or Enhances not in mainstream repositories:
  INLA
Availability using Additional_repositories specification:
  INLA   yes   https://inla.r-inla-download.org/R/stable/
Package has a VignetteBuilder field but no prebuilt vignette index.
Found the following (possibly) invalid URLs:
  URL: https://codecov.io/github/meixichen/SpatialGEV (moved to https://app.codecov.io/github/meixichen/SpatialGEV)
    From: README.md
    Status: 200
    Message: OK
```

```
* checking installed package size ... NOTE
  installed size is 102.1Mb
  sub-directories of 1Mb or more:
    libs  101.5Mb
```

```
* checking top-level files ... NOTE
Non-standard file/directory found at top level:
  ‘SpatialGEV.Rproj’
```

```
* checking files in ‘vignettes’ ... WARNING
Files in the 'vignettes' directory but no files in 'inst/doc':
  ‘SpatialGEV-vignette.Rmd’ ‘SpatialGEV-vignette.html’ ‘references.bib’
  ‘sim-1plot-y-1.png’ ‘sim1-plot-abs-1.png’
  ‘sim1-plot-num-obs-per-loc-1.png’ ‘sim1-pos-plots-1.png’
  ‘sim2-plot-ab-1.png’ ‘sim2-pred-plot-1.png’
  ‘snow-display-grid-data-1.png’ ‘snow-display-raw-data-1.png’
  ‘snow-return-plots-1.png’ ‘vignette-helper-functions.R’

* checking re-building of vignette outputs ... ERROR
Error(s) in re-building vignettes:
  ...
--- re-building ‘SpatialGEV-vignette.Rmd’ using rmarkdown
File figure/sim1-plot-abs-1.png not found in resource path
Error: processing vignette 'SpatialGEV-vignette.Rmd' failed with diagnostics:
pandoc document conversion failed with error 99
--- failed re-building ‘SpatialGEV-vignette.Rmd’
```

linking openjournals/joss-reviews#6878
