### 2024-07-10 meixichen/SpatialGEV#22 @meixichen
Thanks for flagging this! My apologies that I did not check the vignette properly. You are right that a pre-compiled vignette is prone to issues like this. I think I decided to precompile the vignette because it took around 10-20 minutes to build. Do you think it is an acceptable length of time? The old precompiled version (without the covariate section) did pass the CRAN test when I submitted it.
