### 2024-07-10 meixichen/SpatialGEV#22 @fabian-s
> I think I decided to precompile the vignette because it took around 10-20 minutes to build. Do you think it is an acceptable length of time?

Not sure -- CRAN policies just say:

> Checking the package should take as little CPU time as possible, as the CRAN check farm is a very limited resource and there are thousands of packages. Long-running tests and vignette code can be made optional for checking, but do ensure that the checks that are left do exercise all the features of the package. [...]
Examples should run for no more than a few seconds each: they are intended to exemplify to the would-be user how to use the functions in the package.

I think it's fine to use pre-compiled for a CRAN submission, but would not recommend it in your development-version on GH.
