### 2024-07-10 #6878 @vissarion
@fabian-s there are no editorial standard specifically for R packages. The general rule "packaged appropriately according to common community standards for the programming language being used", could mean that `R CMD check` should pass but not necessarily `R CMD check --as-cran`. But still I am not saying that `R CMD check` is a requirement for JOSS acceptance but rather a useful tool to indicate if the submitted software is appropriately packaged.
