### 2024-07-10 #6878 @fabian-s
sure, I know, I also do editor-stuff for JOSS - that's why I said *your* standards ;)

FWIW, the NOTES/WARNINGS/ERROS don't seem that serious to me, the package can be installed in the standard way and it runs, including examples etc.

If it were my call, I would still request to make it CRAN-compliant s.t. the version of the package presented in the paper is more similar/identical to the respective CRAN version.
