### 2024-07-10 #6878 @vissarion
I agree that it would be nice to make it CRAN-compliant. Though I would not like this to be a blocking issue. To my understanding there is not much work towards this goal (and in any case maybe that work will be done anyways to update the current CRAN package). @meixichen do you agree to make this package pass `R CMD check --as-cran` in the course of JOSS submission?
