### 2024-07-10 3c32bf8
Resolve vignette errors and include libraries used in vignette in description
