+++ vignettes/SpatialGEV-vignette.Rmd
@@ -29,7 +29,14 @@ MathJax.Hub.Config({
 \newcommand{\ind}{{\overset{ind}{\sim}}}
 \newcommand{\cov}{{\tx{Cov}}}
 
-
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig_path = "",
+  cache=TRUE
+)
+```
 
 ## Introduction to the GEV-GP Model
 The generalized extreme value (GEV) distribution is often used to analyze sequences of maxima within non-overlapping time periods. An example of this type of data is the monthly maximum rainfall levels recorded over years at a weather station. Since there are typically a large number of weather stations within a country or a state, it is more ideal to have a model that can borrow information from nearby weather stations to increase inference and prediction accuracy. Such spatial information is often pooled using the Gaussian process.
@@ -71,8 +78,7 @@ Details about the approximate posterior inference can be found in @chen-etal21.
 ***SpatialGEV*** depends on the package [***TMB***](https://github.com/kaskr/adcomp) to perform the Laplace approximation. Make sure you have ***TMB*** installed following their [instruction](https://github.com/kaskr/adcomp/wiki/Download) before installing ***SpatialGEV***. Moreover, ***SpatialGEV*** uses several functions from the ***INLA*** package for approximating the Matérn covariance with the SPDE representation and for creating meshes on the spatial domain. If the user would like to use the SPDE approximation, please first install ***INLA***. Since ***INLA*** is not on CRAN, it needs to be downloaded following their instruction [here](https://www.r-inla.org/download-install). 
 
 To install the latest version of ***SpatialGEV***, run the following:
-
-```r
+```{r eval=FALSE}
 devtools::install_github("meixichen/SpatialGEV")
 ```
 
@@ -90,8 +96,7 @@ We now demonstrate how to use this package through a simulation study. The simul
 - `logs`: A length $400$ vector of the true values of log-transformed $s_i, \ i=1,\ldots,400$ at the 400 locations 
 
 `a`, `logb` and `logs` are simulated from Gaussian random fields using the R package ***SpatialExtremes***. Using the simulated GEV parameters, we generate 50 to 70 observed data $\yy_i$ at each location $i, \ i=1,\ldots,400$. 
-
-```r
+```{r sim1-getdata, message=FALSE}
 library(SpatialGEV)
 library(fields) # for plots
 library(evd) # for rgev() and qgev()
@@ -107,8 +112,7 @@ y <- Map(rgev, n=sample(50:70, n_loc, replace=TRUE),
 ```
 
 Spatial variation of $a$, $\log(b)$ and $\log(s)$ can be viewed by plotting them on regular lattices:
-
-```r
+```{r sim1-plot-abs, fig.width=14, fig.height=5, out.width="100%"}
 par(mfrow=c(1,3))
 grid_plot(unique(locs$x), unique(locs$y), matrix(a, ncol=sqrt(n_loc)), 
           "Spatial variation of a")
@@ -118,21 +122,15 @@ grid_plot(unique(locs$x), unique(locs$y), matrix(logs, ncol=sqrt(n_loc)),
           "Spatial variation of log(s)", y_lab="")
 ```
 
-<img src="figure/sim1-plot-abs-1.png" alt="plot of chunk sim1-plot-abs" width="100%" />
-
 Number of observations at each location is shown in the figure below.
-
-```r
+```{r sim1-plot-num-obs-per-loc, fig.width=10, fig.height=5, out.width= "100%"}
 barplot(sapply(y, length), 
[…]
