+++ DESCRIPTION
@@ -36,7 +36,10 @@ Suggests:
     INLA,
     testthat,
     knitr,
-    rmarkdown
+    rmarkdown,
+    fields,
+    maps,
+    dplyr
 Additional_repositories: https://inla.r-inla-download.org/R/stable/
 VignetteBuilder: knitr
 URL: https://github.com/meixichen/SpatialGEV
