### 2024-07-10 meixichen/SpatialGEV#22 @meixichen
Have replaced the pre-compiled vignette with the actual rmd for building the vignette. This passed the CRAN check.
