### 2024-07-10 #6878 @meixichen
@fabian-s @vissarion I agree with you both. I have pushed the new changes on the vignette, which is now no longer pre-compiled. Passed the check for me:
```
checking files in 'vignettes' ... OK
* checking examples ... OK
* checking for unstated dependencies in 'tests' ... OK
* checking tests ... [41s] OK
  Running 'testthat.R' [41s]
* checking for unstated dependencies in vignettes ... OK
* checking package vignettes ... OK
* checking re-building of vignette outputs ... [410s] OK
* checking PDF version of manual ... [21s] OK
* checking HTML version of manual ... OK
* DONE
```
