### 2024-07-10 #6878 @fabian-s
> Passed the check for me:

for me as well
