### 2024-07-10 #6878 @fernandomayer
> Hi @fernandomayer just checking if there are any issues that blocks you to start with your review.

@vissarion Sorry for the delay. I'll be starting my review this week, planning to finalise it mid-next week.
