### 2024-10-02 meixichen/SpatialGEV#23 @fernandomayer
Comments from JOSS review
Ref.: https://github.com/openjournals/joss-reviews/issues/6878

Hi @meixichen. I'll try to concentrate all my comments here. Please feel free to reply with any queries may appear.

### Regarding the package itself

- **Installation**:
  - I tested both CRAN and `remotes::install_github()` and everything went fine.
  - When I try to build and check the package locally (with either `R CMD` and `devtools::check()`, I have a few notes:
    - In both the build and the check process, the creation of the vignette takes a lot of time. I'm aware this was previously discussed here  and I just want to add to it. I often have vignettes that takes lots of time to run, so I adopted [this strategy from rOPenSci](https://ropensci.org/blog/2019/12/08/precompute-vignettes/), and also I use to cache the heavier chunks (and ignore the `*_cache` directory in both `.gitignore` and `.Rbuildignore`). You can see an example [here](https://github.com/EireExtremes/geslaR/tree/main/vignettes). Also, see this section of [R Packages book](https://r-pkgs.org/vignettes.html#sec-vignettes-how-built).
