### 2024-10-02 #6878 @fernandomayer
Hi @vissarion and @meixichen.

Sorry for the long time. I completed my review checklist here and left all my comments in https://github.com/meixichen/SpatialGEV/issues/23. Please let me know if there's anything else I can do.

@vissarion My comments on that issue are more for improving the package and don't prevent the paper from being accepted by JOSS.
