### 2024-10-02 #6878 @vissarion
Thanks @fernandomayer!

@meixichen please let us know if you want to address (some of) the comments of https://github.com/meixichen/SpatialGEV/issues/23 as part of this review or you want to leave them as future work.
