### 2024-10-03 #6878 @meixichen
Thank you @fernandomayer for the helpful suggestions. I will try to address most of them. As for @fernandomayer 's suggestion on adding a conclusion paragraph, @vissarion is it ok with you that we add a couple more sentences? I am asking because it seems like our paper is already considered to be on the lengthy side by JOSS standards.
