+++ joss/paper.bib
@@ -2,9 +2,8 @@
 	title={Fast and Scalable Approximate Inference for Spatial Extreme Value Models}, 
 	author={Chen, M. and Ramezan, R. and Lysy, M.},
 	year={2024},
-	journal={Canadian Journal of Statistics (accepted)},
-	url={https://arxiv.org/abs/2110.07051},
-	doi={10.48550/arXiv.2110.07051}
+	journal = {The Canadian Journal of Statistics},
+	doi={10.1002/cjs.11829}
 }
 
 @article{tierney-kadane86,
