+++ vignettes/references.bib
@@ -1,10 +1,9 @@
-@misc{chen-etal21,
-  url = {https://arxiv.org/abs/2110.07051},
+@article{chen-etal21,
   author = {Chen, Meixi and Ramezan, Reza and Lysy, Martin},
-  keywords = {stat.CO},
-  title = {Fast Approximate Inference for Spatial Extreme Value Models},
-  publisher = {arXiv}, 
-  year = {2022}
+  year = {2024},
+  title = {Fast and Scalable Approximate Inference for Spatial Extreme Value Models},
+  journal = {The Canadian Journal of Statistics},
+  doi={10.1002/cjs.11829}
 }
 
 
