### 2024-10-03 0e67922
changed reference DOI in DESCRIPTION to CJS article
