+++ vignettes/SpatialGEV-vignette.Rmd
@@ -1,9 +1,14 @@
 ---
-title: "A Guide to the ***SpatialGEV*** Package"
-output: rmarkdown::html_vignette
+title: "A Guide to the SpatialGEV Package"
+author: "Meixi Chen, Martin Lysy, Reza Ramezan"
+output: 
+  rmarkdown::html_vignette:
+    toc: true
 bibliography: references.bib
+link-citations: true
+csl: technometrics.csl
 vignette: >
-  %\VignetteIndexEntry{SpatialGEV-guide}
+  %\VignetteIndexEntry{A Guide to the SpatialGEV Package}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
@@ -20,25 +25,32 @@ MathJax.Hub.Config({
 \newcommand{\XX}{{\bm{X}}}
 \newcommand{\YY}{{\bm{Y}}}
 \newcommand{\ZZ}{{\bm{Z}}}
+\newcommand{\bbe}{{\bm{\beta}}}
 \newcommand{\tth}{{\bm{\theta}}}
 \newcommand{\pps}{{\bm{\psi}}}
 \newcommand{\uu}{{\bm{u}}}
+\newcommand{\BB}{{\bm{B}}}
+\newcommand{\TTh}{{\bm{\Theta}}}
 \newcommand{\SSi}{{\bm{\Sigma}}}
 \newcommand{\VV}{{\bm{V}}}
 \newcommand{\iid}{{\overset{iid}{\sim}}}
 \newcommand{\ind}{{\overset{ind}{\sim}}}
-\newcommand{\cov}{{\tx{Cov}}}
+\newcommand{\cov}{{\operatorname{Cov}}}
+\newcommand{\obs}{{\tx{observed}}}
+\newcommand{\normal}{\operatorname{Normal}}
 
 ```{r, include = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>",
   fig_path = "",
-  cache=TRUE
+  cache = TRUE,
+  tidy = "styler"
 )
 ```
 
 ## Introduction to the GEV-GP Model
+
 The generalized extreme value (GEV) distribution is often used to analyze sequences of maxima within non-overlapping time periods. An example of this type of data is the monthly maximum rainfall levels recorded over years at a weather station. Since there are typically a large number of weather stations within a country or a state, it is more ideal to have a model that can borrow information from nearby weather stations to increase inference and prediction accuracy. Such spatial information is often pooled using the Gaussian process.
 
 The GEV-GP model is a hierarchical model with a data layer and a spatial random effects layer. Let $\xx_1, \ldots, \xx_n \in \mathbb{R}^2$ denote the geographical coordinates of $n$ locations, and let $y_{ik}$ denote the extreme value measurement $k$ at location $i$, for $k = 1, \ldots, n_i$. The data layer specifies that each observation $y_{ik}$ has a generalized extreme value distribution, denoted by $y \sim \tx{GEV}(a, b, s)$, whose CDF is given by
@@ -50,37 +62,54 @@ The GEV-GP model is a hierarchical model with a data layer and a spatial random 
     \end{cases} 
 \label{eqn:gev-distn}
 \end{equation} 
-where $a\in\mathbb{R}$, $b>0$, and $s\in\mathbb{R}$ are location, scale, and shape parameters, respectively. The support of the GEV distribution depends on the parameter values: $y$ is bounded below by $a-b/s$ when $s>0$, bounded above by $a-b/s$ when $s<0$, and unbounded when $s=0$. To capture the spatial dependence in the data, we assume some or all of the GEV parameters in the data layer are spatially varying. Thus they are introduced as random effects in the model. 
[…]
