+++ DESCRIPTION
@@ -38,7 +38,8 @@ Suggests:
     rmarkdown,
     fields,
     maps,
-    dplyr
+    dplyr,
+    styler
 Additional_repositories: https://inla.r-inla-download.org/R/stable/
 VignetteBuilder: knitr
 URL: https://github.com/meixichen/SpatialGEV
