### 2024-10-05 23f3d04
Precompiled vignette for CRAN
- Source vignette now lives in `vignettes/articles/_SpatialGEV.Rmd` and
  generates precompiled vignette via `vignettes/articles/precompile.R`.
- `.Rbuildignore`d `vignettes/articles`.
  As a result, copied `vignettes-helper-functions.R` to `vignettes/` so that
  full code to regenerate vignette is available when downloading package from
  CRAN.
- `.gitignore`d cache files.
- Changed vignette file name to `SpatialGEV.Rmd`.
