+++ vignettes/articles/_SpatialGEV.Rmd
@@ -1,0 +1,623 @@
+---
+title: "A Guide to the SpatialGEV Package"
+author: "Meixi Chen, Martin Lysy, Reza Ramezan"
+date: "`r Sys.Date()`"
+output: 
+  rmarkdown::html_vignette:
+    toc: true
+bibliography: references.bib
+link-citations: true
+csl: technometrics.csl
+vignette: >
+  %\VignetteIndexEntry{A Guide to the SpatialGEV Package}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+<script type="text/x-mathjax-config">
+MathJax.Hub.Config({
+  TeX: { equationNumbers: { autoNumber: "AMS" } }
+});
+</script>
+
+\newcommand{\bm}[1]{\boldsymbol{#1}}
+\newcommand{\tx}[1]{\mathrm{#1}}
+\newcommand{\xx}{{\bm{x}}}
+\newcommand{\yy}{{\bm{y}}}
+\newcommand{\XX}{{\bm{X}}}
+\newcommand{\YY}{{\bm{Y}}}
+\newcommand{\ZZ}{{\bm{Z}}}
+\newcommand{\bbe}{{\bm{\beta}}}
+\newcommand{\tth}{{\bm{\theta}}}
+\newcommand{\pps}{{\bm{\psi}}}
+\newcommand{\uu}{{\bm{u}}}
+\newcommand{\BB}{{\bm{B}}}
+\newcommand{\TTh}{{\bm{\Theta}}}
+\newcommand{\SSi}{{\bm{\Sigma}}}
+\newcommand{\VV}{{\bm{V}}}
+\newcommand{\iid}{{\overset{iid}{\sim}}}
+\newcommand{\ind}{{\overset{ind}{\sim}}}
+\newcommand{\cov}{{\operatorname{Cov}}}
+\newcommand{\obs}{{\tx{observed}}}
+\newcommand{\normal}{\operatorname{Normal}}
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig.cap = "",
+  fig.path = "SpatialGEV_figures/",
+  cache.path = "SpatialGEV_cache/",
+  cache = TRUE,
+  tidy = "styler"
+)
+```
+
+## Introduction to the GEV-GP Model
+
+The generalized extreme value (GEV) distribution is often used to analyze sequences of maxima within non-overlapping time periods. An example of this type of data is the monthly maximum rainfall levels recorded over years at a weather station. Since there are typically a large number of weather stations within a country or a state, it is more ideal to have a model that can borrow information from nearby weather stations to increase inference and prediction accuracy. Such spatial information is often pooled using the Gaussian process.
+
+The GEV-GP model is a hierarchical model with a data layer and a spatial random effects layer. Let $\xx_1, \ldots, \xx_n \in \mathbb{R}^2$ denote the geographical coordinates of $n$ locations, and let $y_{ik}$ denote the extreme value measurement $k$ at location $i$, for $k = 1, \ldots, n_i$. The data layer specifies that each observation $y_{ik}$ has a generalized extreme value distribution, denoted by $y \sim \tx{GEV}(a, b, s)$, whose CDF is given by
+\begin{equation}
+    F(y\mid a, b, s) =
+    \begin{cases}
+    \exp\left\{-\left(1+s\frac{y-a}{b}\right)^{-\frac{1}{s}}\right\} \ \ &s\neq 0,\\
+    \exp\left\{-\exp\left(-\frac{y-a}{b}\right)\right\} \ \ &s=0,
+    \end{cases} 
+\label{eqn:gev-distn}
+\end{equation} 
[…]
