+++ vignettes/SpatialGEV.Rmd
@@ -1,6 +1,7 @@
 ---
 title: "A Guide to the SpatialGEV Package"
 author: "Meixi Chen, Martin Lysy, Reza Ramezan"
+date: "2024-10-05"
 output: 
   rmarkdown::html_vignette:
     toc: true
@@ -39,15 +40,7 @@ MathJax.Hub.Config({
 \newcommand{\obs}{{\tx{observed}}}
 \newcommand{\normal}{\operatorname{Normal}}
 
-```{r, include = FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>",
-  fig_path = "",
-  cache = TRUE,
-  tidy = "styler"
-)
-```
+
 
 ## Introduction to the GEV-GP Model
 
@@ -100,6 +93,17 @@ The package provides an interface to estimate the approximate joint posterior di
 
 - `spatialGEV_predict()`: Sample from the posterior predictive distribution $p(\yy(\XX_{\tx{new}}) \mid \XX)$ at new locations $\XX_{\tx{new}}$.
 
+A fundamental application of GEV-GP modeling is the prediction of return levels.  For a given spatial location $\xx$, the return level $p$ is defined as the $(1-p) \times 100\%$ quantile of the GEV distribution at that location:
+\begin{equation}
+\begin{aligned}
+z_p(\xx) & = z_p(a(\xx), b(\xx), s(\xx)) \\
+& = F^{-1}(1-p \mid a(\xx), b(\xx), s(\xx)) \\
+& = a(\xx) - \exp(b(\xx) - s(\xx)) \left\{1 - [-\log(1-p)]^{-\exp(s(\xx))}\right\}.
+\end{aligned}
+\label{eq:return}
+\end{equation}
+Each of the functions above can be used to estimate the posterior distribution of return levels $p(z_p(\XX^\star) \mid \YY)$ at arbitary locations $\XX^\star$ (observed or new).
+
 Details about the approximate posterior inference can be found in @chen-etal21.
 
 ## Installation
@@ -107,12 +111,14 @@ Details about the approximate posterior inference can be found in @chen-etal21.
 ***SpatialGEV*** depends on the package [***TMB***](https://github.com/kaskr/adcomp) to perform the Laplace approximation. Make sure you have ***TMB*** installed as described [here](https://github.com/kaskr/adcomp/wiki/Download) before installing ***SpatialGEV***. Moreover, ***SpatialGEV*** uses several functions from the [***INLA***](https://www.r-inla.org/) package for approximating the Matérn covariance with the SPDE representation and for creating meshes on the spatial domain. If the user would like to use the SPDE approximation (highly recommended for larger datasets), please first install ***INLA***. Since ***INLA*** is not on CRAN, it needs to be installed as described [here](https://www.r-inla.org/download-install). 
 
 Once ***TMB*** and ***INLA*** have been installed, one can install the CRAN version of ***SpatialGEV*** via the following command:
-```{r eval=FALSE}
+
+``` r
 install.packages("SpatialGEV")
 ```
 
 To install the latest version of ***SpatialGEV***, run the following:
-```{r eval=FALSE}
+
+``` r
 devtools::install_github("meixichen/SpatialGEV")
 ```
 
@@ -131,7 +137,8 @@ We now demonstrate how to use this package through a simulation study. The simul
[…]
