### 2024-10-05 8d7bbb2
Add conclusion section and modify bib
