+++ joss/paper.bib
@@ -2,7 +2,7 @@
 	title={Fast and Scalable Approximate Inference for Spatial Extreme Value Models}, 
 	author={Chen, M. and Ramezan, R. and Lysy, M.},
 	year={2024},
-	journal = {The Canadian Journal of Statistics},
+	journal = {Canadian Journal of Statistics (Accepted)},
 	doi={10.1002/cjs.11829}
 }
 
