### 2024-10-07 20d34fd
added plot of marginal posteriors to README
