+++ README.md
@@ -7,7 +7,6 @@
 
 [![](https://www.r-pkg.org/badges/version/SpatialGEV)](https://cran.r-project.org/package=SpatialGEV)
 [![R-CMD-check](https://github.com/meixichen/SpatialGEV/workflows/R-CMD-check/badge.svg)](https://github.com/meixichen/SpatialGEV/actions)
-[![codecov](https://codecov.io/github/meixichen/SpatialGEV/branch/joss/graph/badge.svg?token=O7XZWKV2FD)](https://codecov.io/github/meixichen/SpatialGEV)
 <!-- badges: end -->
 
 *Meixi Chen, Martin Lysy, Reza Ramezan*
@@ -68,33 +67,50 @@ logb <- simulatedData2$logb # log scale
 logs <- simulatedData2$logs # log shape
 locs <- simulatedData2$locs # coordinate matrix
 n_loc <- nrow(locs) # number of locations
-y <- Map(evd::rgev, n=sample(50:70, n_loc, replace=TRUE),
-         loc=a, scale=exp(logb), shape=exp(logs)) # observations
+y <- Map(evd::rgev,
+  n = sample(50:70, n_loc, replace = TRUE),
+  loc = a, scale = exp(logb), shape = exp(logs)
+) # observations
 
-filled.contour(unique(locs$x), unique(locs$y), matrix(a, ncol=sqrt(n_loc)), 
-               color.palette = terrain.colors, xlab="Longitude", ylab="Latitude", 
-               main="Spatial variation of a",
-               cex.lab=1,cex.axis=1)
+filled.contour(
+  x = unique(locs$x),
+  y = unique(locs$y),
+  z = matrix(a, ncol = sqrt(n_loc)),
+  color.palette = terrain.colors,
+  xlab = "Longitude", ylab = "Latitude",
+  main = "Spatial variation of a",
+  cex.lab = 1, cex.axis = 1
+)
 ```
 
 <img src="man/figures/README-show-data-1.png" width="50%" />
 
 ``` r
 
-filled.contour(unique(locs$x), unique(locs$y), matrix(exp(logb), ncol=sqrt(n_loc)), 
-               color.palette = terrain.colors, xlab="Longitude", ylab="Latitude", 
-               main="Spatial variation of b",
-               cex.lab=1,cex.axis=1)
+filled.contour(
+  x = unique(locs$x),
+  y = unique(locs$y),
+  z = matrix(exp(logb), ncol = sqrt(n_loc)),
+  color.palette = terrain.colors,
+  xlab = "Longitude", ylab = "Latitude",
+  main = "Spatial variation of b",
+  cex.lab = 1, cex.axis = 1
+)
 ```
 
 <img src="man/figures/README-show-data-2.png" width="50%" />
 
 ``` r
 
-filled.contour(unique(locs$x), unique(locs$y), matrix(exp(logs), ncol=sqrt(n_loc)),
-               color.palette = terrain.colors, xlab="Longitude", ylab="Latitude",
-               main="Spatial variation of s",
-               cex.lab=1,cex.axis=1)
+filled.contour(
+  x = unique(locs$x),
+  y = unique(locs$y),
+  z = matrix(exp(logs), ncol = sqrt(n_loc)),
+  color.palette = terrain.colors,
+  xlab = "Longitude", ylab = "Latitude",
+  main = "Spatial variation of s",
+  cex.lab = 1, cex.axis = 1
+)
 ```
 
 <img src="man/figures/README-show-data-3.png" width="50%" />
@@ -108,20 +124,26 @@ SPDE-approximated Matérn kernel `kernel="spde"`. Initial parameter
 values are passed to `init_param` using a list.
 
 ``` r
[…]
