### 2024-10-07 fd005ff
changed size of figure in README
