+++ README.md
@@ -143,7 +143,7 @@ fit <- spatialGEV_fit(
 class(fit)
 #> [1] "spatialGEVfit"
 print(fit)
-#> Model fitting took 16.8532750606537 seconds 
+#> Model fitting took 18.7642500400543 seconds 
 #> The model has reached relative convergence 
 #> The model uses a spde kernel 
 #> Number of fixed effects in the model is 9 
@@ -170,18 +170,18 @@ the sample object.
 pos_summary <- summary(sam)
 pos_summary$param_summary[1:5, ]
 #>        2.5%      25%      50%      75%    97.5%     mean
-#> a1 59.25026 60.40500 61.00427 61.62747 62.77069 61.01379
-#> a2 59.48056 60.54463 61.11959 61.68137 62.73596 61.11448
-#> a3 59.48051 60.46909 60.97925 61.49504 62.45936 60.97861
-#> a4 59.21269 60.20313 60.73353 61.25842 62.22433 60.72580
-#> a5 58.79040 59.85486 60.45759 61.05277 62.13024 60.45854
+#> a1 59.42203 60.51864 61.07608 61.65162 62.72965 61.08421
+#> a2 59.51169 60.49154 60.99219 61.53318 62.51913 61.00703
+#> a3 59.51886 60.42647 60.90600 61.40069 62.33711 60.91756
+#> a4 59.41940 60.33537 60.81912 61.30541 62.24544 60.82448
+#> a5 59.35572 60.28673 60.76393 61.25759 62.21742 60.76773
 pos_summary$y_summary[1:5, ]
 #>        2.5%      25%      50%      75%    97.5%     mean
-#> y1 39.40958 55.14766 67.50013 83.91931 139.5049 72.83111
-#> y2 38.93668 55.55349 68.17977 85.45240 139.8687 73.54768
-#> y3 38.40468 55.14562 68.31470 86.16494 143.2262 73.79189
-#> y4 37.37779 54.84899 67.84350 86.21062 141.1534 73.42529
-#> y5 35.48020 54.17413 67.58432 86.12147 140.6352 73.03575
+#> y1 38.07950 55.22167 68.23224 85.55679 137.1879 73.11054
+#> y2 37.69110 55.21894 67.88157 84.81316 135.3150 72.74923
+#> y3 36.65306 54.78220 68.18212 85.56453 138.1629 72.99914
+#> y4 36.19345 54.81739 68.08520 85.55948 141.2051 73.08745
+#> y5 36.31394 54.29342 67.38804 86.26536 143.0015 73.22856
 ```
 
 One can also plot the full posteriors using e.g., the
