### 2024-10-08 38d7fa2
Avoid edge test case where spatial resolution is chosen wrong
