### 2024-10-08 #6878 @meixichen
@vissarion @fernandomayer we've now addressed the majority of @fernandomayer review comments.  Here is a summary of exactly what we did and didn't do:

- **Installation:** We did not make any changes to the package in this regard.  The installation is already being tested via `R cmd check` on multiple OS and release/devel versions of R via a widely used [GitHub action](https://github.com/meixichen/SpatialGEV/blob/joss/.github/workflows/R-CMD-check.yaml), and we've also tested the package on CRAN's own [Windows](https://win-builder.r-project.org/) system (results available [here](https://win-builder.r-project.org/K0xY2C77ho3M/) for the next ~72 hours).

- **Vignette:**
    - We used the suggested method to ship a precompiled vignette.  We realized this goes against @fabian-s's [request](https://github.com/openjournals/joss-reviews/issues/6878#issuecomment-2220129379).  However, it occurs to us that CRAN will not accept a package for which the vignette compile time is more than a few seconds.  In anticipation of uploading the latest version of our package to CRAN if/when the JOSS paper is accepted, we believe that we'll need to shorten the vignette compile time sooner or later during the JOSS review.

        Precompiling the vignette is easily handled using the suggested method via a [`precompile.R`](https://github.com/meixichen/SpatialGEV/blob/joss/vignettes/articles/precompile.R) script.  However we do still need to remember to precompile the vignette any time we make changes to the underlying [source file](https://github.com/meixichen/SpatialGEV/blob/joss/vignettes/articles/_SpatialGEV.Rmd). This could be automated via e.g., a `pre-push` git hook similar to what is done for rendering `Readme.Rmd` to `Readme.md` [here](https://lorenzwalthert.github.io/precommit/articles/available-hooks.html#readme-rmd-rendered).  However, we would like to request to add this functionality at a later stage -- given how easy it is to run `precompile.R` vs how much effort it would take to implement the `pre-push` hook at this stage.

        - `vignettes/vignette-helper-functions.R`: We decided to continue `source()`ing this file rather than adding its contents directly to the package.  The file only contains a few plotting functions which are too specific to be exported, and we believe it to be bad R practice to use non-exported functions in any user-facing code.  We could always dump the contents of the file into the beginning of the vignette, but this seems like a detour which would distract the readers.  So we just put a comment next to the [line](https://github.com/meixichen/SpatialGEV/blob/8d7bbb233818f30866aa15f62db8e991e7d338ef/vignettes/articles/_SpatialGEV.Rmd#L152) where the `source()` command is called, briefly explaining the contents of the file.

        - All other vignette comments are addressed.

- **Readme.Rmd:** We've added a few lines of code to plot the marginal posterior parameter distributions using the [**bayesplot**](https://mc-stan.org/bayesplot/) package.

- **Conclusion:** We've added a few lines to the JOSS manuscript to end the paper less abruptly.
