### 2024-10-21 #6878 @meixichen
Hi @vissarion, the version number is v1.0.2 and the DOI is [DOI: 10.5281/zenodo.13968085](https://doi.org/10.5281/zenodo.13968085)
