### 2024-10-22 f9ac55f
Fix paper doi
