+++ joss/paper.bib
@@ -95,7 +95,8 @@
   booktitle = {The Handbook of {Markov Chain Monte Carlo}},
   publisher = {Chapman \& Hall / CRC Press},
   year = {2011},
-  chapter = {5}
+  chapter = {5},
+  doi = {10.1201/b10905}
 }
 
 @article{hoffman-gelman14,
@@ -105,7 +106,7 @@
   journal = {Journal of Machine Learning Research},
   volume = {15},
   pages = {1593-1623},
-  doi = {10.5555/2627435.2638586}
+  url = {https://dl.acm.org/doi/10.5555/2627435.2638586}
 }
 
  @Misc{rstan,
