### 2024-10-22 #6878 @meixichen
> ⚠️ [Error preparing paper acceptance](https://github.com/openjournals/joss-papers/actions/runs/11455145476?check_suite_focus=true). The generated XML metadata file is invalid.
>
> ```yaml
> Element fig content does not follow the DTD, expecting (object-id* , label* , caption* , abstract* , kwd-group* , subj-group* , (alt-text | long-desc | email | ext-link | uri)* , (disp-formula | disp-formula-group | chem-struct-wrap | disp-quote | speech | statement | verse-group | table-wrap | p | def-list | list | alternatives | array | code | graphic | media | preformat | xref)* , (attrib | permissions)*), got (caption fig fig )
> ```

@vissarion Can you please advise what this error means?
