### 2024-10-22 #6878 @vissarion
@meixichen it seems that the error is related to the subfigures. Maybe it worth trying to merge the two figures into one file and the simply include that file in markdown.
