### 2024-10-23 88f594b
Add missing DOI for R packages and merge subfigures into one png in joss md
