+++ joss/paper.bib
@@ -115,6 +115,7 @@
    note = {R package version 2.21.2},
    year = {2020},
    url = {http://mc-stan.org/},
+   doi = {10.32614/CRAN.package.rstan}
  }
  
 
@@ -167,13 +168,15 @@
    author = {Ribatet, M. and Singleton, R. and {R Core team}},
    year = {2022},
    howpublished = {Version 2.1-0. \url{https://CRAN.R-project.org/package=SpatialExtremes}},
+   doi = {10.32614/CRAN.package.SpatialExtremes}
  }
 
 @manual{mgcv,
-    title = {\texttt{mgcv}: Mixed {GAM} Computation Vehicle with Automatic Smoothness Estimation},
+    title = {mgcv: Mixed {GAM} Computation Vehicle with Automatic Smoothness Estimation},
     author = {Wood, Simon N.},
     year = {2023},
-    howpublished = {Version 1.9-1. \url{https://CRAN.R-project.org/package=mgcv}}
+    howpublished = {Version 1.9-1. \url{https://CRAN.R-project.org/package=mgcv}},
+    doi = {10.32614/CRAN.package.mgcv}
 }
   
 @article{youngman22,
@@ -195,6 +198,7 @@
   number = {2},
   pages = {31-32},
   note = {https://journal.r-project.org/articles/RN-2002-015/},
-  issn = {1609-3631}
+  issn = {1609-3631},
+  doi = {10.32614/CRAN.package.evd}
 }
 
