### 2024-10-23 #6878 @meixichen
Thanks @vissarion

I have fixed the missing and invalid DOIs, and updated the figure in the md file. The new DOI is [10.5281/zenodo.13978363](https://doi.org/10.5281/zenodo.13978363)
