### 2024-10-23 #6878 @vissarion
I updated the version to v3 since that was the version mentioned in zenodo. You have to create a new tagged release with
that version.
