### 2024-10-23 #6878 @vissarion
But you have to change the archive too. I think it is simpler to just change the version in zenodo, it is possible? i.e. v1.0.2
