### 2024-10-23 #6878 @meixichen
Can't seem to changet the version number on zenodo. I changed the GitHub release tag to v3 and updated the associated link of the archive to https://github.com/meixichen/SpatialGEV/tree/v3. Does this work?
