### 2024-10-24 #6878 @vissarion
Regarding skip DOIs in the check, I confirm that https://dl.acm.org/doi/10.5555/2627435.2638586 resolves correctly.
