### 2024-11-08 #6878 @crvernon
@meixichen - thank you for your patience.  I just need you to address the following and I'll move to accept this for publication:

- [ ] The license specified in your Zenodo archive does not match the one that is specified in your code repo.  Please edit the Zenodo metadata to reflect the correct license.  No need to generate another release.

That's it.  This paper is very clean.  Excellent work!
