### 2024-11-08 #6878 @meixichen
Thank you @crvernon! I have modified the license on Zenodo.
