### 2024-11-08 #6878 @crvernon
@meixichen will you remove the special characters around "R" in the title (`\textsf{R}'` just needs to be an R)?  It is causing the title to render odd when posted on JOSS's website.  After you do, I'll reaccept to fix this.
