### 2024-11-08 631e9f1
Remove textsf around R in paper.md
