### 2024-11-08 #6878 @meixichen
@crvernon sure, I made the change in the joss branch (since this paper.md is associated with the joss branch I assume it is ok?)
